requirements.txt
openai>=1.0.0
notion-client>=2.2.1
requests>=2.31.0
python-dotenv>=1.0.0
schedule>=1.2.0
feedparser>=6.0.10
beautifulsoup4>=4.12.0
python-dateutil>=2.8.2